refactor(core): Dedup formatBytes and toMb into @n8n/utils (no-changelog)
