+++ packages/cli/src/modules/data-table/multer-upload-middleware.ts
@@ -2,6 +2,7 @@
 import { Logger, safeJoinPath } from '@n8n/backend-common';
 import { GlobalConfig } from '@n8n/config';
 import { Service } from '@n8n/di';
+import { formatBytes } from '@n8n/utils/number/bytes';
 import type { Request, RequestHandler } from 'express';
 import { mkdir, readdir, stat, unlink } from 'fs/promises';
 import multer from 'multer';
@@ -18,7 +19,6 @@ import {
 	type MulterFilenameCallback,
 	type UploadMiddleware,
 } from './types';
-import { formatBytes } from './utils/size-utils';
 
 const ALLOWED_EXTENSIONS = ['.csv'];
