+++ packages/frontend/editor-ui/src/app/utils/typesUtils.test.ts
@@ -1,10 +1,4 @@
-import {
-	isEmpty,
-	intersection,
-	isValidDate,
-	formatBytes,
-	abbreviateNumber,
-} from '@/app/utils/typesUtils';
+import { isEmpty, intersection, isValidDate, abbreviateNumber } from '@/app/utils/typesUtils';
 
 describe('Types Utils', () => {
 	describe('isEmpty', () => {
@@ -44,22 +38,6 @@ describe('Types Utils', () => {
 			expect(
 				intersection([1, 2, 2, 3, 4], [2, 3, 3, 4], [2, 1, 5, 4, 4, 1], [2, 4, 5, 5, 6, 7]),
 			).toEqual([2, 4]);
-		});
-	});
-
-	describe('formatBytes', () => {
-		test.each([
-			[0, '0B'],
-			[512, '512B'],
-			[1023, '1023B'],
-			[1024, '1KB'],
-			[1536, '2KB'],
-			[100 * 1024 + 44 * 1024, '144KB'],
-			[1024 * 1024 - 1, '1024KB'],
-			[1024 * 1024, '1MB'],
-			[5 * 1024 * 1024, '5MB'],
-		])('formats %i bytes as %s', (input, expected) => {
-			expect(formatBytes(input)).toBe(expected);
 		});
 	});
